ExtensionModuleParser
