+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/parsers/v2/ExtensionModuleParser.java
@@ -2,6 +2,7 @@ package com.sap.cloud.lm.sl.mta.parsers.v2;
 
 import static com.sap.cloud.lm.sl.mta.handlers.v2.Schemas.EXT_MODULE;
 
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
@@ -10,25 +11,32 @@ import java.util.Set;
 import com.sap.cloud.lm.sl.common.ParsingException;
 import com.sap.cloud.lm.sl.common.util.ListUtil;
 import com.sap.cloud.lm.sl.mta.model.v2.ExtensionModule;
+import com.sap.cloud.lm.sl.mta.model.v2.ExtensionModule.Builder;
 import com.sap.cloud.lm.sl.mta.model.v2.ExtensionProvidedDependency;
 import com.sap.cloud.lm.sl.mta.model.v2.ExtensionRequiredDependency;
-import com.sap.cloud.lm.sl.mta.model.v2.ExtensionModule.Builder;
 import com.sap.cloud.lm.sl.mta.parsers.ListParser;
+import com.sap.cloud.lm.sl.mta.parsers.ModelParser;
 import com.sap.cloud.lm.sl.mta.schema.MapElement;
 
-public class ExtensionModuleParser extends com.sap.cloud.lm.sl.mta.parsers.v1.ExtensionModuleParser {
+public class ExtensionModuleParser extends ModelParser<ExtensionModule> {
 
+    protected static final String PROCESSED_OBJECT_NAME = "MTA extension module";
+
+    public static final String NAME = "name";
+    public static final String PROPERTIES = "properties";
+    public static final String PROVIDES = "provides";
     public static final String PARAMETERS = "parameters";
     public static final String REQUIRES = "requires";
 
+    protected Set<String> usedProvidedDependencyNames = Collections.emptySet();
     protected Set<String> usedRequiredDependencyNames = new HashSet<>();
 
     public ExtensionModuleParser(Map<String, Object> source) {
-        super(EXT_MODULE, source);
+        this(EXT_MODULE, source);
     }
 
     protected ExtensionModuleParser(MapElement schema, Map<String, Object> source) {
-        super(schema, source);
+        super(PROCESSED_OBJECT_NAME, schema, source);
     }
 
     @Override
@@ -40,6 +48,14 @@ public class ExtensionModuleParser extends com.sap.cloud.lm.sl.mta.parsers.v1.Ex
         builder.setRequiredDependencies2(getExtensionRequiredDependencies2());
         builder.setProvidedDependencies2(getExtensionProvidedDependencies2());
         return builder.build();
+    }
+
+    protected String getName() throws ParsingException {
+        return getStringElement(NAME);
+    }
+
+    protected Map<String, Object> getProperties() {
+        return getMapElement(PROPERTIES);
     }
 
     protected Map<String, Object> getParameters() {
@@ -57,14 +73,23 @@ public class ExtensionModuleParser extends com.sap.cloud.lm.sl.mta.parsers.v1.Ex
     }
 
     protected List<ExtensionProvidedDependency> getExtensionProvidedDependencies2() {
-        return ListUtil.cast(getExtensionProvidedDependencies1());
+        return ListUtil.cast(getExtensionProvidedDependencies());
+    }
+
+    protected List<ExtensionProvidedDependency> getExtensionProvidedDependencies() {
+        return getListElement(PROVIDES, new ListParser<ExtensionProvidedDependency>() {
+            @Override
+            protected ExtensionProvidedDependency parseItem(Map<String, Object> map) {
+                return getProvidedDependencyParser(map).setUsedValues(usedProvidedDependencyNames)
+                    .parse();
+            }
+        });
     }
 
     protected ExtensionRequiredDependencyParser getRequiredDependencyParser(Map<String, Object> source) {
         return new com.sap.cloud.lm.sl.mta.parsers.v2.ExtensionRequiredDependencyParser(source);
     }
 
-    @Override
     protected ExtensionProvidedDependencyParser getProvidedDependencyParser(Map<String, Object> source) {
         return new com.sap.cloud.lm.sl.mta.parsers.v2.ExtensionProvidedDependencyParser(source);
     }
